requirements.txt
appnope==0.1.3
asn1crypto==1.5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base45==0.4.3
cbor2==5.4.3
certvalidator==0.11.1
cffi==1.15.0
cose==0.9.dev8
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
ecdsa==0.17.0
entrypoints==0.4
executing==0.8.3
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.4
oscrypto==1.3.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.0.0
scipy==1.8.1
six==1.16.0
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
typing_extensions==4.2.0
wcwidth==0.2.5
xgboost==1.6.1

